Add support for multiple jobs
This patch provides support for concurrent builds with Make and Boost
Build with a similar mechanism to the -j parameter. By specifying
--jobs on the command line with an optional number, Needy will provide
the appropriate parameter to Make or Boost Build.

In the event that a particular library can't be build using multiple
jobs, but other libraries can, a maximum number of jobs can be
specified in the configuration for that library using the
'max-concurrency' parameter. See the PostgreSQL example for details.