Allow custom Make prefix arguments
Some projects such as lua use prefix variables that differ from the
established standards. In the case of LUA, the prefix is
"INSTALL_TOP". While it isn't economical to include all possible
prefix variables, it is desirable to be able to provide overrides
during certain parts of the build process.

Ideally, to accomplish this, configuration parameters would be
provided that are able to be passed to the appropriate subsystems.
Before this is possible, however, there needs to be a system of
variable substitution for configurations as this example requires not
only a custom installation argument, but also knowledge of where
exactly it is to be installed.

In the absence of this system, a single configuration parameter,
'make-prefix-arg' has added to support the specific use case of
non-standard prefix flags.