Allow custom project targets
With some projects, it's not desirable to build the entire thing. In
some other projects, a choice of targets is required such as with Lua.

To address this use case, a configuration item, "project-targets", has
been added where an array of targets can be specified. It is up to the
project type to properly make use of those targets if applicable.