Add better logging in some places
In an ongoing effort to improve needy's log output, several variables
and actions are now logged in this patch. In particular,
`process.__log_check_call` previously would not provide output stderr
of called commands. This patch fixes that oversight and now includes
stderr for command of the appropriate verbosity level.