Fix regression with SourceProject
Previously, Needy would only ensure that the include and lib
directories existed in the post-build step. When that was changed to
the prebuild step, SourceProjects.copy_headers would fail due to the
include directory always existing prior to the copy.
distutils.dir_util.copy_tree exhibits the same behavior as
shutil.copytree but won't fail when the destination directory exists.