remove linkage flag
narrowing the scope of needy to be less of a build system abstraction.
if you need a particular linkage, you can get it just as easily via
configure-args, b2-args, or make-target. this has the added benefits of
making needy more transparent and serving as a hint for which build
system should be used.