Add build artifact caching
This patch introduces caching that can be made available with `needy cache set [path]` and acts as a key/value store to archive the contents of the build directory after builds. When the cache contains artifacts for a particular build, they are used instead of going through the build process.