Fix Android compiler selection
Previously, when Androids builds were made, clang was prioritized as the compiler of choice but during a refactor of the Android source, clang was inadvertently removed as an option due to a typo in the os.path.exists call. This patch fixes that issue and raises an exception if a compiler can't be found instead of allowing the compilation to progress to a later stage.