Prevent pkgconfig generation with no libraries or includes (#146)
Previously, when Needy generated pkgconfig files, it would do so indiscriminately as long as a pkgconfig file didn't already exist. This patch ensures that Needy only generates a pkgconfig file when there appears to be build output that can be included or linked to.

Additionally, it fixes issues related to subprojects including tools required for a build that aren't built because Needy discovers the tool to be located in pkgconfig.